+++ packages/dialog/data.js
@@ -1,0 +1,25 @@
+module.exports = {
+  // 标题
+  title: '',
+  // 自定义 btn 列表
+  // { type: 按钮类型，回调时以此作为区分依据，text: 按钮文案, color: 按钮文字颜色 }
+  buttons: [],
+  // 内容
+  message: ' ',
+  // 选择节点
+  selector: '#zan-dialog',
+  // 按钮是否展示为纵向
+  buttonsShowVertical: false,
+  // 是否展示确定
+  showConfirmButton: true,
+  // 确认按钮文案
+  confirmButtonText: '确定',
+  // 确认按钮颜色
+  confirmButtonColor: '#3CC51F',
+  // 是否展示取消
+  showCancelButton: false,
+  // 取消按钮文案
+  cancelButtonText: '取消',
+  // 取消按钮颜色
+  cancelButtonColor: '#333'
+};
